# docs/swagger.yaml
openapi: 3.1.0
...
info:
  title: Argus
  version: "0.0.2"
  summary: Active monitoring for domains.
  description: |
    A node probes the targets in its configuration and exposes what it found.
    Everything under `/api` exists to run and change those checks from outside;
    everything else is what a scraper or a person reads.

    **Permissions.** Reading and writing are separate, and each is configured on
    its own (`http.api.read_auth` / `write_auth`, either `token` or `none`).
    A read token never authorises a change: handing a dashboard the operator's
    token would be handing it the power to repoint the node. `/metrics` and the
    status pages carry no token at all — they hold no secrets, and a scraper has
    none to offer.

    **The API is off by default.** It appears only with `http.api.enabled: true`.
    Without it every path below `/api` is a 404.
servers:
  - url: http://{host}:{port}
    variables:
      host: {default: localhost}
      port: {default: "6767"}
...
paths:
...
  /logo.png:
    get:
      tags: [status]
      summary: The mark the status page wears
      description: |
        Carried in the binary, so a node in a closed network renders the same as
        one with a route to the internet. Served beside the page rather than
        inlined into it, so a browser fetches it once for a page that is meant
        to be reloaded often.
      security: []
      responses:
        "200":
          description: A PNG, about ten kilobytes.
          content:
            image/png:
              schema: {type: string, format: binary}
        "304":
          description: Unchanged since the client last asked.
...
  /probe:
    get:
      tags: [checks]
      summary: Check one target and answer with a Prometheus exposition
      description: |
        The other way round: Prometheus names the target and gets
        the measurements back, so the list of targets lives in its service
        discovery rather than in this node's configuration. Nothing is stored,
        exactly as for `/api/check`.

        Alongside the probe's own metrics it returns `probe_success`, one series
        per target. `probe_up` cannot answer "did the check pass": it is per
        backend, and a target that stopped resolving has none at all, so an
        alert written over it quietly stops firing at the worst moment.

        Series carry no prefix by default, so `probe_success`,
        `probe_duration_seconds` and `probe_up` are the names a blackbox
        dashboard already uses. Alongside them, unless
        `http.api.probe_aliases` is false, the endpoint publishes
        blackbox_exporter's names for the series both tools measure the same
        way — `probe_http_status_code`, `probe_ssl_earliest_cert_expiry`,
        `probe_dns_lookup_time_seconds`, `probe_ip_protocol` and the rest.
        Only exact twins are translated: a familiar name over a different
        measurement is worse than no series at all.

        No histograms: one run is not a distribution, and a histogram scraped
        here has a count of one and never moves. The `*_last_duration_seconds`
        gauge beside it carries the measurement.

        **Permission.** The same rule as `/api/check`: a configured probe run as
        configured is a read; naming `target` or `hostname` is a write, because
        both send this node somewhere the configuration did not say to send it,
        carrying whatever credentials the probe holds. `debug` is a write too:
        the trace carries what the target answered and what was asked of it. A
        scrape config supplies the token with `authorization`.

        Served only when `http.api.probe_endpoint` is true, which is the
        default. Turned off, the path is not mounted at all.
      security: [{readToken: []}]
      parameters:
        - name: probe
          in: query
          required: false
          description: The configured probe to run. `module` is accepted as an alias.
          schema: {type: string}
          example: web
        - name: module
          in: query
          required: false
          description: An alias for `probe`, so an existing scrape config needs no rewriting.
          schema: {type: string}
        - name: target
          in: query
          required: false
          description: Check this target instead of the probe's own. Supplying it makes the request a write.
          schema: {type: string}
          example: https://www.google.com/
        - name: hostname
          in: query
          required: false
          description: |
            Present this name to the target — Host header and SNI — without
            changing the address connected to, which is how one virtual host is
            checked on a machine that serves many. Supplying it makes the
            request a write.
          schema: {type: string}
          example: canary.example.com
        - name: debug
          in: query
          required: false
          description: |
            Prepend the run's trace as comment lines. The answer is still a
            valid exposition, so a scrape is unaffected and whoever is holding
            curl gets the log. Supplying it makes the request a write.
          schema: {type: boolean, default: false}
      responses:
        "200":
          description: |
            The check ran. A target that is down is a successful answer to the
            question asked, so the status describes the request and
            `probe_success` describes the target.
          content:
            text/plain:
              schema: {type: string}
              example: |
                # HELP probe_success Whether the on-demand check of this target passed.
                # TYPE probe_success gauge
                probe_success{probe="web",target="https://www.google.com/"} 1
        "400":
          description: No such probe, or the request did not describe a check.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "405": {$ref: "#/components/responses/MethodNotAllowed"}
  /api/config:
    get:
      tags: [config]
      summary: Read the check configuration
      description: |
        The configuration as accepted, byte for byte: comments, templates and
        defaults are still there and `${VAR}` references are not expanded, so a
        read token sees no secret taken from the environment and what comes back
        can be sent straight back in. A node started from several files answers
        a YAML stream with one document per file, which `PUT` refuses: such a
        configuration is edited on disk. Empty when there is none yet.
      security: [{readToken: []}]
      responses:
        "200":
          description: The configuration, in YAML.
          content:
            application/yaml:
              schema: {type: string}
              example: |
                defaults: {interval: 30s}
                probes:
                  - name: web
                    type: http
                    targets: ["https://www.google.com/"]
                    http: {}
        "401": {$ref: "#/components/responses/Unauthorized"}
    put:
      tags: [config]
      summary: Replace the check configuration
      description: |
        Parsed exactly as a file is — an unknown field is an error, not a probe
        that runs happily while checking the wrong thing, and `${VAR}` is
        expanded — then applied, then written back to the file the node started
        from. What is written is the request body as it arrived: comments and
        `${VAR}` references survive, and a secret taken from the environment
        does not reach the disk.

        The order matters. Building the probes is the rest of the validation, so
        a configuration this node cannot run never reaches the file; and a change
        that cannot be written is refused outright, because applying it in memory
        alone is how a node and its files drift apart.

        Writing needs a single destination: `PUT` replaces the whole
        configuration, and putting that back into a directory would mean deciding
        which of the files already there to delete. A node started with
        `-probes <dir>` or several paths answers 409 and says so in its log at
        startup.

        Probes whose definition did not change keep running, with their tickers
        untouched: a node reloaded often must not stampede its targets.

        `POST` does the same thing.
      security: [{writeToken: []}]
      requestBody:
        required: true
        content:
          application/yaml:
            schema: {type: string}
            example: |
              probes:
                - name: web
                  type: http
                  targets: ["https://www.google.com/"]
                  http: {}
      responses:
        "200":
          description: Applied and written. The body is the node's new state.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Status"}
        "400":
          description: |
            The configuration was refused and nothing changed — a syntax error,
            an unknown field, a probe type this build does not have, a timeout
            longer than its interval. The running checks are untouched.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "409":
          description: |
            This node has no single check file to write to, so the change was
            not applied at all.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
              example:
                error: >-
                  this node has no single check file to write to, so a change
                  through the API would not survive a restart; edit the file and
                  reload instead
        "500":
          description: |
            Applied but not saved. The node is running this configuration and the
            file still holds the previous one, so a restart will undo it. The
            message says so rather than leaving the operator to guess which of
            the two is live.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "405": {$ref: "#/components/responses/MethodNotAllowed"}
    post:
      tags: [config]
      summary: Replace the check configuration
      description: The same as `PUT`.
      security: [{writeToken: []}]
      requestBody:
        required: true
        content:
          application/yaml:
            schema: {type: string}
      responses:
        "200":
          description: Applied and written.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Status"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "409":
          description: No single check file to write to.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "500":
          description: Applied but not saved.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
  /api/reload:
    post:
      tags: [config]
      summary: Re-read the configuration files
      description: |
        What SIGHUP does. The files the node was started with are read again, so
        an edit made on disk takes effect without a restart. Probes that did not
        change keep running.
      security: [{writeToken: []}]
      responses:
        "200":
          description: Reloaded. The body is the node's new state.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Status"}
        "400":
          description: |
            The files were refused, or the node was started with none. Nothing
            changed.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "405": {$ref: "#/components/responses/MethodNotAllowed"}
  /api/probes:
    post:
      tags: [checks]
      summary: Run a probe now
      description: |
        Runs one probe — or every probe, with no `name` — without waiting for its
        interval, and records the result like any other run. This is how a deploy
        verifies a new check.

        Unlike `/api/check`, this one writes to the store: it is a real run,
        brought forward.
      security: [{writeToken: []}]
      parameters:
        - name: name
          in: query
          required: false
          description: The probe to run. Omitted, every probe runs.
          schema: {type: string}
          example: web
      responses:
        "200":
          description: The run finished.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ran: {type: string, description: The probe that ran, or "all".}
                required: [ran]
              example: {ran: web}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404":
          description: No probe of that name is configured on this node.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "409":
          description: |
            That probe is already running. Two runs of one probe would report
            each other's results, so the second is refused rather than queued.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "405": {$ref: "#/components/responses/MethodNotAllowed"}
  /api/check:
    get:
      tags: [checks]
      summary: Check something once, storing nothing
      description: |
        Ask a question and read the answer inline. Nothing reaches the store: a
        one-off check run by a person debugging a target must not leave a series
        behind that a dashboard then treats as monitoring data.

        **The permission depends on what is asked.** Running a configured probe
        exactly as configured is a read. Pointing it at another address
        (`target`) is a write, because it sends that probe's configured
        credentials somewhere the caller chose.
      security: [{readToken: []}]
      parameters:
        - {$ref: "#/components/parameters/CheckProbe"}
        - {$ref: "#/components/parameters/CheckTarget"}
        - {$ref: "#/components/parameters/CheckHostname"}
        - {$ref: "#/components/parameters/CheckDebug"}
      responses:
        "200":
          description: |
            The check ran. A target that is down is a successful answer to the
            question that was asked, so the status describes the API call and
            `success` describes the target.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CheckResult"}
        "400":
          description: No such probe, or the request did not describe a check.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "405": {$ref: "#/components/responses/MethodNotAllowed"}
    post:
      tags: [checks]
      summary: Check a probe definition that is not in the configuration
      description: |
        The body is a probe definition, parsed as strictly as one in a file. This
        is always a write: a definition can name the `external` prober, which
        runs a command on this node.

        With an empty body this behaves exactly like `GET`.
      security: [{writeToken: []}]
      parameters:
        - {$ref: "#/components/parameters/CheckProbe"}
        - {$ref: "#/components/parameters/CheckTarget"}
        - {$ref: "#/components/parameters/CheckHostname"}
        - {$ref: "#/components/parameters/CheckDebug"}
      requestBody:
        required: false
        content:
          application/yaml:
            schema: {type: string}
            example: |
              name: ad-hoc
              type: http
              targets: ["https://www.google.com/"]
              http:
                validators:
                  - {name: status, status: "200-399"}
      responses:
        "200":
          description: The check ran.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CheckResult"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
    put:
      tags: [checks]
      summary: Check a probe definition that is not in the configuration
      description: The same as `POST`.
      security: [{writeToken: []}]
      parameters:
        - {$ref: "#/components/parameters/CheckProbe"}
        - {$ref: "#/components/parameters/CheckTarget"}
        - {$ref: "#/components/parameters/CheckHostname"}
        - {$ref: "#/components/parameters/CheckDebug"}
      requestBody:
        required: false
        content:
          application/yaml:
            schema: {type: string}
      responses:
        "200":
          description: The check ran.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CheckResult"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
components:
  securitySchemes:
    readToken:
      type: http
      scheme: bearer
      description: |
        `http.api.read_token`, or `http.api.token` when no read token is set.
        Not required when `read_auth: none`.
    writeToken:
      type: http
      scheme: bearer
      description: |
        `http.api.token`. A read token is never accepted here. Not required when
        `write_auth: none`, which the node warns about at startup.
  parameters:
    CheckProbe:
      name: probe
      in: query
      required: false
      description: A configured probe to run. Leave it out when the body carries a definition.
      schema: {type: string}
      example: web
    CheckTarget:
      name: target
      in: query
      required: false
      description: |
        Check this one target instead of the probe's own. Supplying it makes the
        request a write.
      schema: {type: string}
      example: https://staging.google.com/
    CheckHostname:
      name: hostname
      in: query
      required: false
      description: |
        Present this name to the target — Host header and SNI — without changing
        the address connected to. Supplying it makes the request a write.
      schema: {type: string}
      example: canary.example.com
...
  responses:
    Unauthorized:
      description: The token is missing or does not grant this action.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
          example: {error: unauthorized}
    BadRequest:
      description: The request was refused and nothing changed.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    MethodNotAllowed:
      description: Wrong method for this path.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: |
            What went wrong, in a sentence. Failures name the thing that failed —
            the field, the probe, the file — because the next step after reading
            this is fixing it.
      required: [error]
    Status:
      type: object
      description: What the node currently checks. No metrics.
      properties:
        node: {type: string, description: This node's name, which every series carries as a label.}
        version: {type: string, example: 0.0.2, description: The release this node runs, for a deploy to gate on.}
        uptime: {type: string, example: 3h12m4s}
        series: {type: integer, description: Series held in the metric store.}
        config_source:
          type: string
          enum: [file, api, none]
          description: >
            Where the running checks came from. "none" is a node nobody has
            pushed to yet: it checks nothing and says so.
        config_hash:
          type: string
          description: >
            sha256 of the bytes the configuration was accepted as. Whoever
            pushes can compute the same value and skip a push that would change
            nothing. A configuration read from files hashes over their names and
            contents instead, so it never matches a pushed document.
        config_applied_at: {type: string, format: date-time}
        resolver: {type: string, enum: [system, custom]}
        probes:
          type: array
          items: {$ref: "#/components/schemas/ProbeStatus"}
      required: [node, version, uptime, series, config_source, resolver, probes]
...